get_atno changed: return array of ints instead of floats64

+++ PyQuante/Molecule.py
@@ -135,7 +135,7 @@ class Molecule:
         atno = [] 
         for atom in self.atoms: 
             atno.append(atom.atno)
-        atno = array(atno,float64) 
+        atno = array(atno,int) 
         return atno
 
     def get_atms(self):
